Remove Categories, setCurrentCategory via viewModel
